Tests at least compile
