+++ na-client/test/client/client-test.js
@@ -1,0 +1,12 @@
+import { EventEmitter } from 'events';
+import path from 'path';
+import binaryjs from 'binaryjs';
+
+import { expect } from 'chai';
+import sinon from 'sinon';
+
+import Client from '../../js/client/client';
+
+describe('Client', function() {
+
+});
