+++ na-client/test/files/file-watcher-test.js
@@ -6,103 +6,9 @@ import {expect} from 'chai';
 import path from 'path';
 import sinon from 'sinon';
 
-import {buildFileInfoForBackend} from '../../../js/lib/files/files-parsing';
-import {setupFilePathProcessor, setupFileWatcher} from '../../../js/lib/files/files-processing';
+import fileWatcher from '../../js/files/file-watcher';
 
-describe('setupFilePathProcessor', function() {
-    let sendFileData;
-    let processTracks;
-    let musicDir;
-    let updateHash;
-    let digestHash;
-    let hash;
-
-    const key = 'aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa';
-    const mockHostedTracks = {
-        'Genre': {
-            'Artist': {
-                'Album': {
-                    '01': {
-                        title: 'Song Title',
-                        id: 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
-                    }
-                }
-            }
-        }
-    };
-
-
-    beforeEach(function() {
-        sendFileData = sinon.spy();
-        processTracks = sinon.stub().returns(mockHostedTracks);
-        musicDir = ['home', 'music'].join(path.sep);
-        digestHash = {
-            digest: sinon.stub().returns('xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx')
-        };
-        updateHash = {
-            update: sinon.stub().returns(digestHash)
-        };
-        hash = sinon.stub(crypto, 'createHash').returns(updateHash);
-    });
-
-    afterEach(function() {
-        sendFileData = null;
-        processTracks = null;
-        crypto.createHash.restore();
-    });
-
-    it('should return a function', function() {
-        const results = setupFilePathProcessor(sendFileData, processTracks, musicDir, key);
-
-        expect(results).to.be.a('function');
-    });
-
-    it('should create processor that does not send data on err', function() {
-        const processFiles = setupFilePathProcessor(sendFileData, processTracks, musicDir, key);
-
-        processFiles('Error', []);
-
-        expect(sendFileData.called).to.be.false;
-        expect(processTracks.called).to.be.false;
-    });
-
-    it('should create processor that processes files and sends data', function() {
-        const processFiles = setupFilePathProcessor(sendFileData, processTracks, musicDir, key);
-        const files = [['home', 'music', 'Genre', 'Artist', 'Album', '01-Song Title.mp3'].join(path.sep)];
-
-        processFiles(null, files);
-
-        const expectedTracks = [{
-            genre: 'Genre',
-            artist: 'Artist',
-            album: 'Album',
-            file: ['home', 'music', 'Genre', 'Artist', 'Album', '01-Song Title.mp3'].join(path.sep),
-            number: '01',
-            title: 'Song Title',
-            id: 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
-        }];
-
-        expect(processTracks.calledWith(expectedTracks, sinon.match.func, {})).to.be.true;
-        expect(sendFileData.calledWith({
-            key: key,
-            tracks: {
-                'Genre': {
-                    'Artist': {
-                        'Album': {
-                            '01': {
-                                title: 'Song Title',
-                                id: 'xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx'
-                            }
-                        }
-                    }
-                }
-            }
-        })).to.be.true;
-        console.log(sendFileData.firstCall.args);
-    })
-});
-
-describe('setupFileWatcher', function() {
+describe('fileWatcher', function() {
     let sendFileData;
     let processTracks;
     let musicDir;
