+++ na-client/test/streams/stream-test.js
@@ -1,0 +1,11 @@
+import path from 'path';
+import { EventEmitter } from 'events';
+
+import { expect } from 'chai';
+import sinon from 'sinon';
+
+import Stream from '../../js/streams/stream';
+
+describe('Stream', function() {
+
+});
